+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-_nothing changed yet_
+### Removed
+
+- Unneccessary internal constructor (for UnitTests)
 
 
 ## [v0.1.0] - 2024-06-21